+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## Unreleased
 
-- Add request context to `HttpException` and `SocketException`
-- `SocketException` and `FileSystemException` with `OSError`s report the `OSError` as root exception
+- Add request context to `HttpException` and `SocketException` ([#1118](https://github.com/getsentry/sentry-dart/pull/1118))
+- `SocketException` and `FileSystemException` with `OSError`s report the `OSError` as root exception ([#1118](https://github.com/getsentry/sentry-dart/pull/1118))
 ### Features
 
 - Feat: Screenshot Attachment ([#1088](https://github.com/getsentry/sentry-dart/pull/1088))